feat(secret): expose internal methods as a library
