+++ packages/secret/README.md
@@ -2,15 +2,38 @@
 
 ![npm](https://img.shields.io/npm/v/@node-cli/secret?label=version&logo=npm)
 
-> Secret is a command line tool that can encode or decode a file with a password.
+> Secret is a command line tool that can encode or decode a file with a password. It also exposes a library to be used in other projects.
 
-## Installation
+## Installation as a CLI
 
 ```sh
 > npm install --global @node-cli/secret
 ```
 
-## Examples
+## Installation as a library
+
+```sh
+> cd your-project
+> npm install @node-cli/secret
+```
+
+## Usage as a library
+
+### Encrypt a string with a password
+
+```js
+import { encrypt } from "@node-cli/secret";
+const encrypted = encrypt("password", "Hello World");
+```
+
+### Decrypt a string with a password
+
+```js
+import { decrypt } from "@node-cli/secret";
+const decrypted = decrypt("password", encrypted);
+```
+
+## Usage as a CLI
 
 NOTE: The password is not stored anywhere, it is only used to encrypt or decrypt the file and will be prompted each time.
