+++ packages/secret/package.json
@@ -6,11 +6,9 @@
 	"description": "Secret is a CLI tool that can encode or decode a file with a password",
 	"type": "module",
 	"types": "./dist/index.d.ts",
-	"exports": "./dist/secret.js",
+	"exports": "./dist/lib.js",
 	"bin": "dist/secret.js",
-	"files": [
-		"dist"
-	],
+	"files": ["dist"],
 	"node": ">=16",
 	"scripts": {
 		"build": "npm-run-all --serial clean build:types build:js build:barrel",
@@ -21,6 +19,7 @@
 		"lint": "biome lint src",
 		"test": "cross-env-shell NODE_OPTIONS=--experimental-vm-modules TZ=UTC jest",
 		"test:coverage": "npm run test -- --coverage",
+		"test:watch": "npm run test -- --watch",
 		"watch": "swc --strip-leading-paths --watch --out-dir dist src"
 	},
 	"dependencies": {
